requirements.txt
scrapling>=0.2.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
psutil>=5.9.0

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
responses>=0.23.0
freezegun>=1.2.0